requirements.txt
django
django-extensions
django-mptt
django-mptt-admin
